backend_full/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.3.0
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
pydantic-settings==2.12.0
python-multipart==0.0.21